extract github cli commands
